test(runtime): share the LabelCoreRepro ceiling fixture and recalibrate #783's ceilings
Independent review of #783's ceiling tests found two problems:

1. issue_783_refine_memory_ceiling.rs duplicated LabelCoreRepro (already
   copied identically in issue_730_bfs_memory_ceiling.rs and
   issue_730_explicit_memory_ceiling.rs) and its self-violating variant a
   second time, worsening the exact "same fixture in 2+ places" duplication
   the project's implementation policy forbids -- "share it" was the #783
   task's own explicit instruction, and this had drifted from it further,
   not converged.
2. That new copy also silently dropped `invariant`/`trans`/`reachable`
   declarations present in the canonical fixture, while its own doc comment
   claimed it was "identical". That is not cosmetic: it changes the
   KernelModel's clone size, which is exactly what #783's pre-fix
   measurements are calibrated against.

Fix: extracted `LABEL_CORE_REPRO_SOURCE` and `Fixture` into
`tests/support/mod.rs` (the same `#[path = "support/mod.rs"] mod support;`
pattern `rust/fslc/tests/support/mod.rs` already uses), and all three
`fsl-runtime` ceiling test files now import it -- the two existing files'
own copies are retired along with the new one, so this drift class cannot
recur in any of the three by construction. `rust/fslc/tests/issue_697_all_properties_memory.rs`
carries a fourth copy in a different crate a `tests/`-local module cannot
reach; consolidating that one is left as an explicit follow-up, not folded
into this fix's scope (a shared dev-dependency crate would be the shape,
and is worth its own issue).

The self-violating variant (base fixture + one `overflow` action) and its
matching all-stutter mapping are now generated as a diff against the
canonical sources (`support::insert_before_closing_brace`) instead of a
second full copy of each.

Recalibration: adding the invariants/trans/reachables back increased the
model's own size, which matters for the *pre-fix* (whole-Monitor-clone)
measurement but not the post-fix one (a `State`-only frontier's memory
barely depends on the static model's size at all) -- exactly the
sensitivity the reviewer flagged. Remeasured (release, `/usr/bin/time -l`
peak memory footprint, corrected fixture):
- check_refinement, depth 3: post-removal ~491 MB (unchanged from the prior,
  drifted-fixture measurement); pre-removal (commit 6012c00) ~1,728 MB (up
  from the drifted fixture's ~1,557 MB, and now an even closer match to
  issue #783's own reported ~1.72 GB). CEILING_KB = 950 * 1024 keeps valid,
  in fact slightly wider, margins (~2.03x above post-removal, ~1.73x below
  pre-removal) and is left unchanged.
- first_self_violation, depth 4: current fix ~492 MB (unchanged); the
  pre-#776 per-node clone, temporarily reintroduced for this measurement
  only, ~2,101 MB (up from ~1,926 MB). CEILING_KB = 1000 * 1024 keeps valid
  margins (~2.03x / ~2.10x) and is left unchanged.

Both ceiling constants happen to need no numeric change; only the cited
measurements in each test's calibration comment are corrected to match the
now-canonical fixture.

Verified end-to-end on macOS (real compile + real run, not just the
`#[cfg(target_os = "linux")]`-gated no-op): temporarily widened the cfg
gate to include macOS and, separately, temporarily bypassed the `ulimit -v`
wrapper (which fails outright on macOS with "Invalid argument" independent
of any of this) so the generated fixtures and shared-module wiring could
actually run; both tests passed against the real (uncapped) binary, both
temporary changes reverted (diffed byte-identical against the intended
content afterward).
